# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: V-Lab Solicitações API
  description: API de gerenciamento de solicitações de atendimento — CIn/UFPE V-Lab
  version: '1.0.0'
servers:
  - url: http://localhost:8000/api/v1
    description: Servidor local de desenvolvimento
security:
  - operatorSession: []
...
paths:
...
  /solicitacoes:
    get:
      tags: [solicitacoes]
      summary: Listar solicitações (paginado)
      description: >-
        Retorna a fila operacional ordenada por estados ativos, prioridade
        (URGENTE, ALTA, MEDIA, BAIXA), data de criação ascendente e id ascendente.
        Sem status_grupo, estados CONCLUIDA e CANCELADA aparecem depois dos estados ativos.
      operationId: listarSolicitacoes
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/Status'
          description: >-
            Filtrar por status. Quando o status resolvido é exatamente AGENDADA
            (este filtro ou data_agendada), a ordenação passa a ser por
            agendado_para e prioridade como desempate, em vez da ordenação por
            prioridade/tempo de espera da fila operacional.
        - name: categoria
          in: query
          schema:
            $ref: '#/components/schemas/Categoria'
          description: Filtrar por categoria
        - name: prioridade
          in: query
          schema:
            $ref: '#/components/schemas/Prioridade'
          description: Filtrar por prioridade
        - name: q
          in: query
          schema: { type: string, maxLength: 100 }
          description: Busca por protocolo ou nome do solicitante.
        - name: data_de
          in: query
          schema: { type: string, format: date }
          description: Limite inicial inclusivo no fuso operacional.
        - name: data_ate
          in: query
          schema: { type: string, format: date }
          description: Limite final inclusivo no fuso operacional.
        - name: ordenar_por
          in: query
          schema: { type: string, enum: [prioridade, data, horario] }
          description: Ordenação no servidor; horario exige resultados agendados.
        - name: direcao
          in: query
          schema: { type: string, enum: [asc, desc] }
          description: Direção aplicável a data e horario.
        - name: status_grupo
          in: query
          schema:
            type: string
            enum: [aberto, encerrado]
          description: >-
            Agrupamento de status usado pela fila e pelo histórico. 'aberto' filtra
            RECEBIDA, EM_ANALISE e AGENDADA; 'encerrado' filtra CONCLUIDA e CANCELADA.
            Ignorado se 'status' também for informado.
        - name: data_agendada
          in: query
          schema:
            type: string
            format: date
            example: '2026-09-25'
          description: >-
            Agenda diária. Dia operacional (fuso AGENDAMENTO_TIMEZONE, padrão America/Recife)
            no formato AAAA-MM-DD. Restringe a AGENDADA cujo agendamento ativo caia nesse dia,
            por dois critérios unidos por OR: (1) modalidade HORARIO com agendado_para no
            intervalo UTC semiaberto [início do dia local, início do dia local seguinte); ou
            (2) modalidade TURNO com data_agendada igual ao dia pedido (TURNO nunca preenche
            agendado_para, por isso o segundo critério é necessário para essas solicitações
            aparecerem na agenda). Ordena por agendado_para (HORARIO primeiro), depois turno
            (MANHA, TARDE, NOITE) para os itens sem horário exato, depois prioridade (URGENTE,
            ALTA, MEDIA, BAIXA), protocolo e id. Combinar com status diferente de AGENDADA ou
            com status_grupo=encerrado retorna 422. Solicitações diferentes no mesmo horário
            são permitidas; igualdade de horário não é conflito.
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 15
      responses:
        '200':
          description: Lista paginada de solicitações
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitacoesPaginadas'
        '422':
          $ref: '#/components/responses/ValidationError'

    post:
      tags: [solicitacoes]
      summary: Criar solicitação
      operationId: criarSolicitacao
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarSolicitacaoBody'
      responses:
        '201':
          description: Solicitação criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitacaoResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /solicitacoes/{id}:
    get:
      tags: [solicitacoes]
      summary: Buscar solicitação por ID
      operationId: buscarSolicitacao
      parameters:
        - $ref: '#/components/parameters/SolicitacaoId'
      responses:
        '200':
          description: Solicitação encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitacaoResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [solicitacoes]
      summary: Editar dados cadastrais da solicitação
      description: >-
        Extensão fora do fluxo obrigatório do edital (que define apenas criar,
        listar/consultar, filtrar e atualizar status). Permite corrigir dados
        cadastrais de uma solicitação que ainda não está em estado final.
      operationId: atualizarSolicitacao
      parameters:
        - $ref: '#/components/parameters/SolicitacaoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarSolicitacaoBody'
      responses:
        '200':
          description: Solicitação atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitacaoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Solicitação em estado final (CONCLUIDA ou CANCELADA) não pode ser editada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [solicitacoes]
      summary: Apagar solicitação definitivamente
      description: >-
        Extensão fora do fluxo obrigatório do edital. Remove o registro
        (hard delete), permitido em qualquer estado.
      operationId: apagarSolicitacao
      parameters:
        - $ref: '#/components/parameters/SolicitacaoId'
      responses:
        '204':
          description: Solicitação apagada com sucesso (sem corpo)
        '404':
          $ref: '#/components/responses/NotFound'
  /solicitacoes/{id}/status:
    patch:
      tags: [solicitacoes]
      summary: Atualizar status da solicitação
      description: >-
        Única via para mudar o status. Para AGENDADA (a partir de EM_ANALISE), é
        obrigatório enviar data_agendada e exatamente um de hora_agendada (modalidade
        HORARIO) ou turno (modalidade TURNO) — nunca os dois, nunca nenhum; interpretados
        no fuso operacional. Para qualquer outro destino esses campos são proibidos (422).
        O instante/turno deve ser estritamente futuro. CONCLUIDA e CANCELADA preservam
        agendado_para (somente para agendamentos HORARIO).
      operationId: atualizarStatus
      parameters:
        - $ref: '#/components/parameters/SolicitacaoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/Status'
                data_agendada:
                  type: string
                  format: date
                  example: '2026-09-25'
                  description: Obrigatória somente quando status=AGENDADA; proibida nos demais.
                hora_agendada:
                  type: string
                  pattern: '^([01]\d|2[0-3]):[0-5]\d$'
                  example: '14:30'
                  description: >-
                    Exclusiva com turno. Obrigatória somente quando status=AGENDADA e a
                    modalidade escolhida é HORARIO; proibida nos demais casos.
                turno:
                  allOf:
                    - $ref: '#/components/schemas/Turno'
                  description: >-
                    Exclusiva com hora_agendada. Obrigatória somente quando status=AGENDADA e
                    a modalidade escolhida é TURNO; proibida nos demais casos.
      responses:
        '200':
          description: Status atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitacaoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/InvalidTransition'
        '422':
          $ref: '#/components/responses/ValidationError'
  /solicitacoes/{id}/agendamento:
    patch:
      tags: [solicitacoes]
      summary: Reagendar solicitação
      description: >-
        Altera somente agendado_para de uma solicitação já AGENDADA; nunca muda o
        status. Reenviar o mesmo instante é idempotente (200, sem escrita). Horários
        iguais aos de outras solicitações são aceitos. A chegada ao local não é registrada.
      operationId: reagendarSolicitacao
      parameters:
        - $ref: '#/components/parameters/SolicitacaoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendamentoInput'
      responses:
        '200':
          description: Agendamento atualizado (ou inalterado, se o instante for o mesmo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitacaoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A solicitação não está AGENDADA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /faltas:
    get:
      tags: [faltas]
      summary: Listar agendamentos em falta
      description: >-
        Agendamentos com status=FALTA, com paciente (telefone mascarado), protocolo,
        data/hora ou turno original e a última tentativa de contato, se houver.
      operationId: listarFaltas
      parameters:
        - name: q
          in: query
          schema: { type: string, maxLength: 100 }
          description: Busca por protocolo ou nome do solicitante.
        - name: prioridade
          in: query
          schema: { $ref: '#/components/schemas/Prioridade' }
        - name: data_de
          in: query
          schema: { type: string, format: date }
        - name: data_ate
          in: query
          schema: { type: string, format: date }
        - name: resultado_contato
          in: query
          schema:
            $ref: '#/components/schemas/ResultadoContato'
        - name: ordenar_por
          in: query
          schema: { type: string, enum: [prioridade, data, horario] }
        - name: direcao
          in: query
          schema: { type: string, enum: [asc, desc] }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 15 }
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'

        '200':
          description: Faltas paginadas
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agendamento'
                  meta:
                    type: object
                    properties:
                      total:        { type: integer }
                      per_page:     { type: integer }
                      current_page: { type: integer }
                      last_page:    { type: integer }
  /agendamentos/{id}/falta:
    post:
      tags: [faltas]
      summary: Registrar falta em um agendamento
      description: >-
        Marca o agendamento AGENDADO como FALTA. solicitacoes.status permanece AGENDADA
        — a falta é só um estado do agendamento. Só é aceito após o horário marcado
        (HORARIO) ou após o fim do turno (TURNO); antes disso retorna 422.
      operationId: registrarFalta
      parameters:
        - $ref: '#/components/parameters/AgendamentoId'
      responses:
        '200':
          description: Falta registrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Agendamento'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Agendamento não está AGENDADO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Ainda não passou do horário marcado ou do fim do turno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /agendamentos/{id}/tentativas-contato:
    post:
      tags: [faltas]
      summary: Registrar tentativa de contato após falta
      operationId: registrarTentativaContato
      parameters:
        - $ref: '#/components/parameters/AgendamentoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrarContatoBody'
      responses:
        '201':
          description: Tentativa registrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TentativaContato'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Agendamento não está em FALTA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/ValidationError'
  /agendamentos/{id}/reagendar-apos-falta:
    post:
      tags: [faltas]
      summary: Reagendar após falta
      description: >-
        Cria um novo Agendamento AGENDADO para a mesma solicitação. Preserva a falta
        anterior como histórico e registra resultado_em nela para impedir novo uso
        dessa falta. Exige solicitação AGENDADA sem agendamento ativo ou posterior.
      operationId: reagendarAposFalta
      parameters:
        - $ref: '#/components/parameters/AgendamentoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgendamentoInput'
      responses:
        '201':
          description: Novo agendamento criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Agendamento'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Solicitação não está AGENDADA, agendamento não está em FALTA, ou já existe um AGENDADO
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  securitySchemes:
    operatorSession:
      type: apiKey
      in: cookie
      name: vlab_session
      description: Laravel session cookie; default name for APP_NAME=VLab. Mutating requests also require the X-XSRF-TOKEN header.
  parameters:
    AgendamentoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: ID do agendamento
    SolicitacaoId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: ID da solicitação
  schemas:
...
    Status:
      type: string
      enum: [RECEBIDA, EM_ANALISE, AGENDADA, CONCLUIDA, CANCELADA]
    Categoria:
      type: string
      enum: [CONSULTA, EXAME, VACINACAO, OUTRO]
    Prioridade:
      type: string
      enum: [BAIXA, MEDIA, ALTA, URGENTE]
    Solicitacao:
      type: object
      properties:
        id:
          type: integer
        protocolo:
          type: string
          example: SOL-2026-0001
        nome_solicitante:
          type: string
        cpf_solicitante:
          type: string
          example: '123.456.789-00'
        data_nascimento:
          type: string
          format: date
        categoria:
          $ref: '#/components/schemas/Categoria'
        prioridade:
          $ref: '#/components/schemas/Prioridade'
        status:
          $ref: '#/components/schemas/Status'
        agendado_para:
          type: string
          format: date-time
          nullable: true
          description: >-
            Instante do atendimento em UTC. Nunca nulo em AGENDADA; sempre nulo em
            RECEBIDA e EM_ANALISE; em CONCLUIDA/CANCELADA guarda o horário histórico
            ou é nulo (cancelada antes de agendar, ou dado legado).
        descricao:
          type: string
        justificativa_prioridade:
          type: string
          nullable: true
        data_criacao:
          type: string
          format: date-time
        data_atualizacao:
          type: string
          format: date-time
        paciente:
          type: object
          allOf:
            - $ref: '#/components/schemas/PacienteResumo'
          nullable: true
        agendamento_ativo:
          type: object
          allOf:
            - $ref: '#/components/schemas/Agendamento'
          nullable: true
    Turno:
      type: string
      enum: [MANHA, TARDE, NOITE]
    ModalidadeAgendamento:
      type: string
      enum: [HORARIO, TURNO]
    StatusAgendamento:
      type: string
      enum: [AGENDADO, REALIZADO, FALTA, CANCELADO]
    ResultadoContato:
      type: string
      enum: [SEM_RESPOSTA, RECADO, CONFIRMOU_RETORNO, NUMERO_INVALIDO]
    AgendamentoInput:
      description: >-
        Exatamente um dos dois formatos — hora exata (modalidade HORARIO) ou
        turno (modalidade TURNO). additionalProperties: false em cada variante
        impede enviar os dois ao mesmo tempo.
      oneOf:
        - type: object
          additionalProperties: false
          required: [data_agendada, hora_agendada]
          properties:
            data_agendada:
              type: string
              format: date
              example: '2026-09-25'
            hora_agendada:
              type: string
              pattern: '^([01]\d|2[0-3]):[0-5]\d$'
              example: '14:30'
        - type: object
          additionalProperties: false
          required: [data_agendada, turno]
          properties:
            data_agendada:
              type: string
              format: date
              example: '2026-09-25'
            turno:
              $ref: '#/components/schemas/Turno'
    PacienteResumo:
      type: object
      properties:
        id: { type: integer }
        nome: { type: string }
        celular_mascarado:
          type: string
          nullable: true
          example: '(81) *****-0000'
          description: Últimos 4 dígitos visíveis; o número completo nunca é retornado pela API.
    TentativaContato:
      type: object
      properties:
        id: { type: integer }
        resultado:
          $ref: '#/components/schemas/ResultadoContato'
        realizada_em:
          type: string
          format: date-time
    Agendamento:
      type: object
      properties:
        id: { type: integer }
        modalidade:
          $ref: '#/components/schemas/ModalidadeAgendamento'
        data_agendada:
          type: string
          format: date
        hora_agendada:
          type: string
          nullable: true
          example: '14:30'
        turno:
          type: string
          allOf:
            - $ref: '#/components/schemas/Turno'
          nullable: true
          description: >-
            Sempre preenchido pelo backend nas duas modalidades (derivado automaticamente
            de hora_agendada quando modalidade=HORARIO); nullable aqui é só defensivo de
            contrato de API, não reflete um caso real em que o valor fica nulo no banco.
        status:
          $ref: '#/components/schemas/StatusAgendamento'
        resultado_em:
          type: string
          format: date-time
          nullable: true
          description: >-
            Momento em que o agendamento foi encerrado ou em que a falta histórica
            foi resolvida por reagendamento.
        falta_registrada_em:
          type: string
          format: date-time
          nullable: true
        falta_corrigida_em:
          type: string
          format: date-time
          nullable: true
          description: Preenchido quando a solicitação é concluída após uma falta registrada por engano.
        solicitacao:
          type: object
          nullable: true
          properties:
            id: { type: integer }
            protocolo: { type: string }
            nome_solicitante: { type: string }
            prioridade:
              $ref: '#/components/schemas/Prioridade'
            status:
              $ref: '#/components/schemas/Status'
            agendamento_ativo:
              type: object
              allOf:
                - $ref: '#/components/schemas/Agendamento'
              nullable: true
            paciente:
              $ref: '#/components/schemas/PacienteResumo'
        ultima_tentativa_contato:
          type: object
          allOf:
            - $ref: '#/components/schemas/TentativaContato'
          nullable: true
    RegistrarContatoBody:
      type: object
      additionalProperties: false
      required: [resultado]
      properties:
        resultado:
          $ref: '#/components/schemas/ResultadoContato'
      description: Payload enxuto — apenas o resultado fechado, sem campo de texto livre.
...
    SolicitacaoResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Solicitacao'
    SolicitacoesPaginadas:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Solicitacao'
        meta:
          type: object
          properties:
            total:       { type: integer }
            per_page:    { type: integer }
            current_page: { type: integer }
            last_page:   { type: integer }
        links:
          type: object
          properties:
            first: { type: string, nullable: true }
            last:  { type: string, nullable: true }
            next:  { type: string, nullable: true }
            prev:  { type: string, nullable: true }
    CriarSolicitacaoBody:
      type: object
      required: [nome_solicitante, cpf_solicitante, data_nascimento, categoria, prioridade, descricao]
      properties:
        nome_solicitante:
          type: string
          minLength: 3
        cpf_solicitante:
          type: string
          pattern: '^\d{3}\.\d{3}\.\d{3}-\d{2}$'
          maxLength: 14
          example: '123.456.789-00'
        data_nascimento:
          type: string
          format: date
          description: Formato AAAA-MM-DD. Deve ser anterior à data atual.
        categoria:
          $ref: '#/components/schemas/Categoria'
        prioridade:
          $ref: '#/components/schemas/Prioridade'
        descricao:
          type: string
          maxLength: 2000
        justificativa_prioridade:
          type: string
          description: Obrigatório quando prioridade=URGENTE
...
    ErrorEnvelope:
      type: object
      properties:
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
  responses:
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ValidationError:
      description: Erro de validação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InvalidTransition:
      description: Transição de status inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'